pyproject.toml
[project]
name = "mpfutils"
version = "0.1.0"
description = "Utilities used in projects by the MeyerPerin Foundation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-storage-blob>=12.24.1",
    "openai>=1.63.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
